修改后台代码

+++ js/background.js
@@ -1,6 +1,8 @@
 (function() {
 
-  var prevTime = 0;
+  setInterval(function() {
+    checkAll();
+  }, 60 * 1000);
 
   chrome.extension.onMessage.addListener(function(request, sender, sendResponse) {
     switch (request.message) {
@@ -23,53 +25,64 @@
       case 'openOptions':
         chrome.extension.getBackgroundPage().open('options.html');
         break;
-      case 'checkRelease':
-        if (request.time - prevTime > 10 * 60 * 1000) {
-          checkAll(request.time);
-        }
-        break;
-      case 'checkReleaseOfReminder':
-        checkRelease(request.reminder, false);
-        var reminder = getReminder(request.reminder.id);
-        sendResponse(reminder);
-        break;
     }
   });
 
-  function checkAll(requestTime) {
-    prevTime = requestTime;
+  chrome.runtime.onConnect.addListener(function(conn) {
+    console.assert(conn.name == "reminder");
+    conn.onMessage.addListener(function(request) {
+      if (request.name == 'search') {
+        search(conn, request);
+      } else if (request.name == 'checkReleaseOfReminder') {
+        checkRelease(request.reminder, function() {
+          var reminder = getReminder(request.reminder.id);
+          conn.postMessage({name: 'checkReleaseOfReminder', reminder: reminder});
+        });
+      }
+    })
+  })
+
+  function checkAll() {
+    console.log('checkRelease');
     var reminderList = getReminderList();
     for (var i = 0; i < reminderList.length; i++) {
       var reminder = reminderList[i];
-      checkRelease(reminderList[i], true);
+      checkRelease(reminderList[i]);
     }
   }
 
-  function checkRelease(reminder, async) {
-    $.ajax({
-      url: reminder.url,
-      type: 'get',
-      async: async,
-      success: function(doc) {
-        var $doc = $(doc);
-        var items = [];
-        $doc.find('.download-list ul li').each(function(_, li) {
-          var text = $(li).text();
-          var url = $(li).find('a').attr('href');
-          if (text.match(/第\d+集/)) {
-            items.push({title: text, url: url, time: formatDate(new Date()) });
-          }
-        })
-
-        var oldItems = reminder.items ? reminder.items : [];
-        if (async && oldItems.length != items.length) {
-          showNotify(reminder.name + "更新啦！\n" + items[items.length-1].title);
-        }
-
-        var newItems = items.slice(0, items.length - oldItems.length);
-        reminder.items = newItems.concat(oldItems);
-        saveReminder(reminder);
+  function checkRelease(reminder, callback) {
+    var engines = [PaoFanEngine, Msj1Engine];
+    var engine = null;
+    $.each(engines, function(_, en) {
+      if (en.match(reminder.name)) {
+        engine = en;
       }
+    });
+    engine.checkRelease(reminder, function(items) {
+      var oldItems = reminder.items ? reminder.items : [];
+      if (!callback && oldItems.length != items.length) {
+        showNotify(reminder.name + "更新啦！\n" + items[items.length-1].title);
+      }
+
+      var newItems = items.slice(0, items.length - oldItems.length);
+      reminder.items = newItems.concat(oldItems);
+      saveReminder(reminder);
+      if (callback) {
+        callback.call(null);
+      }
+    });
+  }
+
+  function search(conn, request) {
+    var keyWord = request.keyWord;
+    var engines = [PaoFanEngine, Msj1Engine];
+    var responseTimes = 0;
+    $.each(engines, function(_, engine) {
+      engine.search(keyWord, function(items) {
+        responseTimes++;
+        conn.postMessage({name: 'search', items: items, finished: responseTimes == engines.length});
+      });
     });
   }
 
@@ -108,6 +121,9 @@
   }
 
   function saveReminder(reminder) {
+    if (!reminder.items && hasExist(reminder.url)) {
+      return;
+    }
     reminder.id = reminder.id ? reminder.id : assignId();
     localStorage.setItem('reminder-' + reminder.id, JSON.stringify(reminder));
     return reminder;
@@ -120,8 +136,8 @@
   function hasExist(url){
     var list = getReminderList();
     for(var i=0;i<list.length;i++){
-      var item = list[i];
-      if(item.url == url){
+      var reminder = list[i];
+      if(reminder.url == url){
         return true;
       }
     }
@@ -139,28 +155,6 @@
       localStorage.setItem(name, maxId);
       return maxId;
     }
-  }
-
-  function formatDate(date) {
-    var dateStr = date.getFullYear() + '-';
-    dateStr += formatInteger(date.getMonth() + 1, 2) + '-';
-    dateStr += formatInteger(date.getDate(), 2) + ' ';
-    dateStr += formatInteger(date.getHours(), 2) + ':';
-    dateStr += formatInteger(date.getMinutes(), 2) + ':';
-    dateStr += formatInteger(date.getSeconds(), 2);
-    return dateStr;
-  }
-
-  function formatInteger(value, length) {
-    var zeroStr = '';
-    var str = value.toString();
-    if (str.length < length) {
-      var zeroNum = length - str.length;
-      for(var i=0;i<zeroNum;i++){
-        zeroStr += '0';
-      }
-    }
-    return zeroStr + str;
   }
 
 })();
